% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{AppSimrel}
\alias{AppSimrel}
\title{Simulation of Multivariate Linear Model Data}
\usage{
AppSimrel()
}
\value{
No return value, runs the shiny interface for simulation
}
\description{
Simulation of Multivariate Linear Model Data
}
