% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesc.r
\name{cov_xy}
\alias{cov_xy}
\title{Covariance between X and Y}
\usage{
cov_xy(obj, use_population = TRUE)
}
\arguments{
\item{obj}{A simrel object}

\item{use_population}{A boolean to specify wheather to use population or sample}
}
\value{
A covariance matrix of X and Y
}
\description{
Covariance between X and Y
}
\keyword{intgernal}
