% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotextra.R
\name{cov_zw}
\alias{cov_zw}
\title{Covariance between Z and W}
\usage{
cov_zw(obj)
}
\arguments{
\item{obj}{A simrel object}
}
\value{
A covariance matrix of Z and W
}
\description{
Helper Functions
}
\keyword{intgernal}
