% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesc.r
\name{cov_zy}
\alias{cov_zy}
\title{Covariance between Z and Y}
\usage{
cov_zy(obj, use_population = TRUE)
}
\arguments{
\item{obj}{A simrel object}

\item{use_population}{A boolean to specify wheather to use population or sample}
}
\value{
A covariance matrix of Z and Y
}
\description{
Covariance between Z and Y
}
\keyword{intgernal}
