% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simulatr.r
\name{plot_simulatr}
\alias{plot_simulatr}
\title{Simulation Plot: The true beta, relevant component and eigen structure}
\usage{
plot_simulatr(obj, ncomp = min(obj$p, obj$n, 20), ask = TRUE,
  print.cov = FALSE, which = 1L:3L)
}
\arguments{
\item{obj}{A simrel object}

\item{ncomp}{Number of components to plot}

\item{ask}{logical, TRUE: functions ask for comfirmation FALSE: function layout plot on predefined format}

\item{print.cov}{Output estimated covariance structure}

\item{which}{A character indicating which plot you want as output, it can take \code{TrueBeta}, \code{RelComp} and \code{EstRelComp}}
}
\value{
A list of plots
}
\description{
Simulation Plot: The true beta, relevant component and eigen structure
}
\keyword{plot}
\keyword{simrel-plot,}
\keyword{simulation}
