\name{simrel-package}
\alias{simrel-package}
\docType{package}
\title{
Package for data simulation and design of computer experiments
}
\description{
Facilitates data simulation from a multiple linear model where 
the data properties can be controlled by a few input parameters. The data 
simulation is based on the concept of relevant latent components and relevant 
predictors, and was developed for the purpose of testing methods for variable 
selection for prediction. The package also contains the tools for designing 
computer experiments in order to investigate the effects of the data properties 
on the performance of the tested methods. The design is constructed using the 
multi-level binary replacement (MBR) design approach which makes it possible to set up fractional designs for 
multi-factor problems with potentially many levels for each factor.}
\details{
\tabular{ll}{
Package: \tab simrel\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2014-11-28\cr
License: \tab GPL-2\cr
}

}
\author{
Solve Sæbø
Maintainer: Solve Sæbø <solve.sabo@nmbu.no>
}

\keyword{ package }

