% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrel-2.r
\name{simrel2}
\alias{simrel2}
\title{Simulation of Multivariate Linear Model data with response}
\usage{
simrel2(n = 50, p = 100, q = c(10, 10, 5), rho = c(0.8, 0.4),
  relpos = list(c(1, 2), c(2, 3)), gamma = 0.5, R2 = c(0.8, 0.8),
  ntest = NULL, muY = NULL, muX = NULL, sim = NULL)
}
\arguments{
\item{n}{Number of training samples}

\item{p}{Number of x-variables}

\item{q}{Vector of number of relevant predictor variables for first, second and common to both responses}

\item{rho}{A 2-element vector, unconditional and conditional correlation between y_1 and y_2}

\item{relpos}{A list of position of relevant component for predictor variables. The list contains vectors of position index, one vector or each response}

\item{gamma}{A declining (decaying) factor of eigen value of predictors (X). Higher the value of \code{gamma}, the decrease of eigenvalues will be steeper}

\item{R2}{Vector of coefficient of determination for each response}

\item{ntest}{Number of test observation}

\item{muY}{Vector of average (mean) for each response variable}

\item{muX}{Vector of average (mean) for each predictor variable}

\item{sim}{A simrel object for reusing parameters setting}
}
\description{
Simulation of Multivariate Linear Model data with response
}
\references{
S<U+00E6>b<U+00F8>, S., Alm<U+00F8>y, T., & Helland, I. S. (2015). simrel<U+2014>A versatile tool for linear model data simulation based on the concept of a relevant subspace and relevant predictors. Chemometrics and Intelligent Laboratory Systems, 146, 128-135.

Alm<U+00F8>y, T. (1996). A simulation study on comparison of prediction methods when only a few components are relevant. Computational statistics & data analysis, 21(1), 87-107.
}
\keyword{data}
\keyword{linear}
\keyword{model}
\keyword{model,}
\keyword{simulation,}

