% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrel.R
\name{simulater}
\alias{simulater}
\title{Multi-Response Simulation of Linear Model Data}
\usage{
simulater(type, ...)
}
\arguments{
\item{type}{Number of training samples}

\item{...}{All required arguments for different types of simulation}
}
\description{
Simulates multivariate linear model data where users can specify few parameters for simulating data with wide range of properties.
}
