% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simulatr.r
\name{simulator_plot}
\alias{simulator_plot}
\title{Simulation Plot: The true beta, relevant component and eigen structure}
\usage{
simulator_plot(obj, ncomp = min(obj$p, obj$n, 20), ask = TRUE,
  print.cov = FALSE, which = c("TrueBeta", "RelComp", "EstRelComp"))
}
\arguments{
\item{obj}{A simrel object}

\item{ncomp}{Number of components to plot}

\item{ask}{logical, TRUE: functions ask for comfirmation FALSE: function layout plot on predefined format}

\item{print.cov}{Output estimated covariance structure}

\item{which}{A character indicating which plot you want as output, it can take \code{TrueBeta}, \code{RelComp} and \code{EstRelComp}}
}
\value{
A list of plots
}
\description{
Simulation Plot: The true beta, relevant component and eigen structure
}
\keyword{plot}
\keyword{simrel-plot,}
\keyword{simulation}

