% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesc.r
\name{tidy_lambda}
\alias{tidy_lambda}
\title{Extract Eigenvalues of predictors}
\usage{
tidy_lambda(obj, use_population = TRUE)
}
\arguments{
\item{obj}{A simrel Object}

\item{use_population}{A boolean to specify where to use population value or calculate from sample}
}
\value{
A dataframe of eigenvalues for each predictors
}
\description{
Extract Eigenvalues of predictors
}
\examples{
sobj <- multisimrel()
sobj \%>\% 
    tidy_lambda()
}
