% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotextra.R
\name{tidy_sigma}
\alias{tidy_sigma}
\title{Tidy covariance matrix}
\usage{
tidy_sigma(covs)
}
\arguments{
\item{covs}{A sigma matrix obtained from cov_mat function}
}
\value{
A tibble with three columns: Predictor, Response and Covariance
}
\description{
Tidy covariance matrix
}
\examples{
sobj <- multisimrel()
head(sigma_df)
}
