% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr.r
\name{validate_param}
\alias{validate_param}
\title{Validating parameter that is passed into \code{simulatr} function}
\usage{
validate_param(par_list)
}
\arguments{
\item{par_list}{A list of parameters that \code{simulatr} takes}
}
\value{
A list of validation message either stop message \code{stop_msg} or warning message \code{warn_msg}
}
\description{
Validating parameter that is passed into \code{simulatr} function
}

