% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesc.r
\name{parse_parm}
\alias{parse_parm}
\title{Some helper function for simulation}
\usage{
parse_parm(character_string, in_list = FALSE)
}
\arguments{
\item{character_string}{A character string for parameter where the items in a list is separated by semicolon. For example: 1, 2; 3, 4}

\item{in_list}{TRUE if the result need to wrap in a list, default is FALSE}
}
\value{
A list or a vector
}
\description{
These function helps to parse a character string into a list object and also creates parameters for performing multiple simulations
}
\examples{
parse_parm("1, 2; 3, 4")
parse_parm("1, 2")
}
