% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrel.r
\name{simrel}
\alias{simrel}
\title{Simulation of Multivariate Linear Model Data}
\usage{
simrel(n, p, q, relpos, gamma, R2, type = "univariate", ...)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of variables.}

\item{q}{Number of predictors related to each relevant components
An integer for univariate, a vector of 3 integers for bivariate and 3 or
more for multivariate simulation (for details see Notes).}

\item{relpos}{A list (vector in case of univariate simulation) of position
of relevant component for predictor variables corresponding
to each response.}

\item{gamma}{A declining (decaying) factor of eigenvalues of predictors (X).
Higher the value of \code{gamma}, the decrease of eigenvalues
will be steeper.}

\item{R2}{Vector of coefficient of determination (proportion of variation
explained by predictor variable) for each relevant response components.}

\item{type}{Type of simulation - \code{univariate}, \code{bivariate} and
\code{multivariate}}

\item{...}{Since this is a wrapper function to simulate univariate,
bivariate or multivariate, it calls their respective function.
This parameter should contain all the necessary arguements for respective
simulations. See \code{\link{unisimrel}}, \code{\link{bisimrel}} and
\code{\link{multisimrel}}}
}
\value{
A simrel object with all the input arguments along with 
    following additional items. For more detail on the return values see the
    individual simulation functions \code{\link{unisimrel}}, 
    \code{\link{bisimrel}} and \code{\link{multisimrel}}.

\strong{Common returns from univariate, bivariate and multivariate simulation:}

\item{call}{the matched call}
\item{X}{simulated predictors}
\item{Y}{simulated responses}
\item{beta}{true regression coefficients}
\item{beta0}{true regression intercept}
\item{relpred}{position of relevant predictors}
\item{n}{number of observations}
\item{p}{number of predictors (as supplied in the arguments)}
\item{p}{number of responses (as supplied in the arguments)}
\item{q}{number of relevant predictors (as supplied in the arguments)}
\item{gamma}{declining factor of eigenvalues of predictors 
    (as supplied in the arguments)}
\item{lambda}{eigenvalues corresponding to the predictors}
\item{R2}{theoretical R-squared value (as supplied in the arguments)}
\item{relpos}{position of relevant components (as supplied in the arguments)}
\item{minerror}{minimum model error}
\item{Sigma}{variance-Covariance matrix of response and predictors}
\item{testX}{simulated test predictor (in univarite simulation \code{TESTX})}
\item{testY}{simulated test response (in univarite simulation \code{TESTY})}
\item{Rotation}{Random rotation matrix used to rotate latent components. Is 
    equivalent to the transpose of eigenvector-matrix. In multivariate
    simulation, \code{Xrotation} (R) and \code{Yrotation} (Q) refers to this matrix
    corresponding to the predictor and response.}
\item{type}{type of simrel object \code{univariate}, \code{bivariate} or 
    \emph{multivariate}}

\strong{Returns from multivariate simulation:}

\item{eta}{a declining factor of eigenvalues of response (Y)
    (as supplied in the arguments)}
\item{ntest}{number of simulated test observations}
\item{W}{simulated response components}
\item{Z}{simulated predictor components}
\item{testW}{test predictor components}
\item{testZ}{test response components}
\item{SigmaWZ}{Variance-Covariance matrix of components of response and predictors}
\item{SigmaWX}{Covariance matrix of response components and predictors}
\item{SigmaYZ}{Covariance matrix of response and predictor components}
\item{RsqW}{Coefficient of determination corresponding to response components}
\item{RsqY}{Coefficient of determination corresponding to response variables}
}
\description{
Simulation of Multivariate Linear Model Data
}
\note{
The parameter \code{q} represetns the number of predictor variables 
   that forms a basis for each of the relevant componetns. For example,
   for \code{q = 8} and relevant components 1, 2, and 3 specified by 
   parameter \code{relpos} then the randomly selected 8 predictor variables 
   forms basis for these three relevant componets and thus in the model 
   these 8 predictors will be revant for the response (outcome).
}
\references{
Sæbø, S., Almøy, T., & Helland, I. S. (2015). simrel—A versatile tool for linear model data simulation based on the concept of a relevant subspace and relevant predictors. Chemometrics and Intelligent Laboratory Systems, 146, 128-135.

Almøy, T. (1996). A simulation study on comparison of prediction methods when only a few components are relevant. Computational statistics & data analysis, 21(1), 87-107.
}
\concept{linear model data}
\concept{simulation}
\keyword{datagen}
